+++ src/include/EnvVars.hpp
@@ -29,7 +29,7 @@ THE SOFTWARE.
 #include "Compatibility.hpp"
 #include "Kernels.hpp"
 
-#define TB_VERSION "1.44"
+#define TB_VERSION "1.45"
 
 extern char const MemTypeStr[];
 extern char const ExeTypeStr[];
@@ -57,9 +57,9 @@ class EnvVars
 {
 public:
   // Default configuration values
-  int const DEFAULT_NUM_WARMUPS          =  3;
-  int const DEFAULT_NUM_ITERATIONS       = 10;
-  int const DEFAULT_SAMPLING_FACTOR      =  1;
+  int const DEFAULT_NUM_WARMUPS       =  3;
+  int const DEFAULT_NUM_ITERATIONS    = 10;
+  int const DEFAULT_SAMPLING_FACTOR   =  1;
 
   // Peer-to-peer Benchmark preset defaults
   int const DEFAULT_P2P_NUM_CPU_SE    = 4;
@@ -75,11 +75,14 @@ public:
 
   // Environment variables
   int alwaysValidate;    // Validate after each iteration instead of once after all iterations
-  int blockSize;         // Size of each threadblock (must be multiple of 64)
-  int blockBytes;        // Each CU, except the last, gets a multiple of this many bytes to copy
+  int blockBytes;        // Each subexecutor, except the last, gets a multiple of this many bytes to copy
   int blockOrder;        // How blocks are ordered in single-stream mode (0=Sequential, 1=Interleaved, 2=Random)
   int byteOffset;        // Byte-offset for memory allocations
   int continueOnError;   // Continue tests even after mismatch detected
+  int gfxBlockSize;      // Size of each threadblock (must be multiple of 64)
+  int gfxSingleTeam;     // Team all subExecutors across the data array
+  int gfxUnroll;         // GFX-kernel unroll factor
+  int gfxWaveOrder;      // GFX-kernel wavefront ordering
   int hideEnv;           // Skip printing environment variable
   int numCpuDevices;     // Number of CPU devices to use (defaults to # NUMA nodes detected)
   int numGpuDevices;     // Number of GPU devices to use (defaults to # HIP devices detected)
@@ -123,10 +126,10 @@ public:
 
   // Enviroment variables only for A2A preset
   int a2aDirect;         // Only execute on links that are directly connected
+  int a2aMode;           // Perform 0=copy, 1=read only, 2 = write only
 
   // Developer features
   int enableDebug;       // Enable debug output
-  int gpuKernel;         // Which GPU kernel to use
 
   // Used to track current configuration mode
   ConfigModeEnum configMode;
@@ -167,19 +170,22 @@ public:
 
     // Different hardware pick different GPU kernels
     // This performance difference is generally only noticable when executing fewer CUs
-    int defaultGpuKernel = 0;
-    if      (archName == "gfx906") defaultGpuKernel = 13;
-    else if (archName == "gfx90a") defaultGpuKernel = 9;
-    else if (archName == "gfx940") defaultGpuKernel = 6;
-    else if (archName == "gfx941") defaultGpuKernel = 6;
-    else if (archName == "gfx942") defaultGpuKernel = 3;
+    int defaultGfxUnroll = 4;
+    if      (archName == "gfx906") defaultGfxUnroll = 13;
+    else if (archName == "gfx90a") defaultGfxUnroll = 9;
+    else if (archName == "gfx940") defaultGfxUnroll = 6;
+    else if (archName == "gfx941") defaultGfxUnroll = 6;
+    else if (archName == "gfx942") defaultGfxUnroll = 4;
 
     alwaysValidate    = GetEnvVar("ALWAYS_VALIDATE"     , 0);
-    blockSize         = GetEnvVar("BLOCK_SIZE"          , 256);
     blockBytes        = GetEnvVar("BLOCK_BYTES"         , 256);
     blockOrder        = GetEnvVar("BLOCK_ORDER"         , 0);
     byteOffset        = GetEnvVar("BYTE_OFFSET"         , 0);
     continueOnError   = GetEnvVar("CONTINUE_ON_ERROR"   , 0);
+    gfxBlockSize      = GetEnvVar("GFX_BLOCK_SIZE"      , 256);
+    gfxSingleTeam     = GetEnvVar("GFX_SINGLE_TEAM"     , 0);
+    gfxUnroll         = GetEnvVar("GFX_UNROLL"          , defaultGfxUnroll);
+    gfxWaveOrder      = GetEnvVar("GFX_WAVE_ORDER"      , 0);
     hideEnv           = GetEnvVar("HIDE_ENV"            , 0);
     numCpuDevices     = GetEnvVar("NUM_CPU_DEVICES"     , numDetectedCpus);
     numGpuDevices     = GetEnvVar("NUM_GPU_DEVICES"     , numDetectedGpus);
@@ -196,7 +202,6 @@ public:
     useXccFilter      = GetEnvVar("USE_XCC_FILTER"      , 0);
     validateDirect    = GetEnvVar("VALIDATE_DIRECT"     , 0);
     enableDebug       = GetEnvVar("DEBUG"               , 0);
-    gpuKernel         = GetEnvVar("GPU_KERNEL"          , defaultGpuKernel);
 
     // P2P Benchmark related
     useDmaCopy        = GetEnvVar("USE_GPU_DMA"         , 0); // Needed for numGpuSubExec
@@ -221,6 +226,7 @@ public:
 
     // A2A Benchmark related
     a2aDirect         = GetEnvVar("A2A_DIRECT"          , 1);
+    a2aMode           = GetEnvVar("A2A_MODE"            , 0);
 
     // Determine random seed
     char *sweepSeedStr = getenv("SWEEP_SEED");
@@ -401,14 +407,14 @@ public:
       printf("[ERROR] Number of GPUs to use (%d) cannot exceed number of detected GPUs (%d)\n", numGpuDevices, numDetectedGpus);
       exit(1);
     }
-    if (blockSize % 64)
+    if (gfxBlockSize % 64)
     {
-      printf("[ERROR] BLOCK_SIZE (%d) must be a multiple of 64\n", blockSize);
+      printf("[ERROR] GFX_BLOCK_SIZE (%d) must be a multiple of 64\n", gfxBlockSize);
       exit(1);
     }
-    if (blockSize > MAX_BLOCKSIZE)
+    if (gfxBlockSize > MAX_BLOCKSIZE)
     {
-      printf("[ERROR] BLOCK_SIZE (%d) must be less than %d\n", blockSize, MAX_BLOCKSIZE);
+      printf("[ERROR] BLOCK_SIZE (%d) must be less than %d\n", gfxBlockSize, MAX_BLOCKSIZE);
       exit(1);
     }
     if (byteOffset % sizeof(float))
@@ -494,9 +500,22 @@ public:
         exit(1);
       }
     }
-    if (gpuKernel < 0 || gpuKernel > NUM_GPU_KERNELS)
+
+    if (a2aMode < 0 || a2aMode > 2)
     {
-      printf("[ERROR] GPU kernel must be between 0 and %d\n", NUM_GPU_KERNELS);
+      printf("[ERROR] a2aMode must be between 0 and 2\n");
+      exit(1);
+    }
+
+    if (gfxUnroll < 1 || gfxUnroll > MAX_UNROLL)
+    {
+      printf("[ERROR] GFX kernel unroll factor must be between 1 and %d\n", MAX_UNROLL);
+      exit(1);
+    }
+
+    if (gfxWaveOrder < 0 || gfxWaveOrder >= 6)
+    {
+      printf("[ERROR] GFX wave order must be between 0 and 5\n");
       exit(1);
     }
 
@@ -533,6 +552,12 @@ public:
       exit(1);
     }
 
+    if (getenv("GPU_KERNEL"))
+    {
+      printf("[WARN] GPU_KERNEL has been deprecated and replaced by GFX_KERNEL and GFX_UNROLL\n");
+      exit(1);
+    }
+
     char* enableSdma = getenv("HSA_ENABLE_SDMA");
     if (enableSdma && !strcmp(enableSdma, "0"))
     {
@@ -553,6 +578,9 @@ public:
     printf(" CONTINUE_ON_ERROR      - Continue tests even after mismatch detected\n");
     printf(" CU_MASK                - CU mask for streams specified in hex digits (0-0,a-f,A-F)\n");
     printf(" FILL_PATTERN=STR       - Fill input buffer with pattern specified in hex digits (0-9,a-f,A-F).  Must be even number of digits, (byte-level big-endian)\n");
+    printf(" GFX_UNROLL             - Unroll factor for GFX kernel (0=auto), must be less than %d\n", MAX_UNROLL);
+    printf(" GFX_SINGLE_TEAM        - Have subexecutors work together on full array instead of working on individual disjoint subarrays\n");
+    printf(" GFX_WAVE_ORDER         - Stride pattern for GFX kernel (0=UWC,1=UCW,2=WUC,3=WCU,4=CUW,5=CWU)\n");
     printf(" HIDE_ENV               - Hide environment variable value listing\n");
     printf(" NUM_CPU_DEVICES=X      - Restrict number of CPUs to X.  May not be greater than # detected NUMA nodes\n");
     printf(" NUM_GPU_DEVICES=X      - Restrict number of GPUs to X.  May not be greater than # detected HIP devices\n");
@@ -589,10 +617,9 @@ public:
     else if (!hideEnv)
       printf("EnvVar,Value,Description,(TransferBench v%s)\n", TB_VERSION);
     if (hideEnv) return;
+
     PRINT_EV("ALWAYS_VALIDATE", alwaysValidate,
              std::string("Validating after ") + (alwaysValidate ? "each iteration" : "all iterations"));
-    PRINT_EV("BLOCK_SIZE", blockSize,
-             std::string("Threadblock size of " + std::to_string(blockSize)));
     PRINT_EV("BLOCK_BYTES", blockBytes,
              std::string("Each CU gets a multiple of " + std::to_string(blockBytes) + " bytes to copy"));
     PRINT_EV("BLOCK_ORDER", blockOrder,
@@ -607,8 +634,20 @@ public:
              (cuMask.size() ? GetCuMaskDesc() : "All"));
     PRINT_EV("FILL_PATTERN", getenv("FILL_PATTERN") ? 1 : 0,
              (fillPattern.size() ? std::string(getenv("FILL_PATTERN")) : PrepSrcValueString()));
-    PRINT_EV("GPU_KERNEL", gpuKernel,
-             std::string("Using GPU kernel ") + std::to_string(gpuKernel) + " [" + std::string(GpuKernelNames[gpuKernel]) + "]");
+    PRINT_EV("GFX_BLOCK_SIZE", gfxBlockSize,
+             std::string("Threadblock size of " + std::to_string(gfxBlockSize)));
+    PRINT_EV("GFX_SINGLE_TEAM", gfxSingleTeam,
+             (gfxSingleTeam ? std::string("Combining CUs to work across entire data array") :
+                              std::string("Each CUs operates on its own disjoint subarray")));
+    PRINT_EV("GFX_UNROLL", gfxUnroll,
+             std::string("Using GFX unroll factor of ") + std::to_string(gfxUnroll));
+    PRINT_EV("GFX_WAVE_ORDER", gfxWaveOrder,
+             std::string("Using GFX wave ordering of ") + std::string((gfxWaveOrder == 0 ? "Unroll,Wavefront,CU" :
+                                                                       gfxWaveOrder == 1 ? "Unroll,CU,Wavefront" :
+                                                                       gfxWaveOrder == 2 ? "Wavefront,Unroll,CU" :
+                                                                       gfxWaveOrder == 3 ? "Wavefront,CU,Unroll" :
+                                                                       gfxWaveOrder == 4 ? "CU,Unroll,Wavefront" :
+                                                                                           "CU,Wavefront,Unroll")));
     PRINT_EV("NUM_CPU_DEVICES", numCpuDevices,
              std::string("Using ") + std::to_string(numCpuDevices) + " CPU devices");
     PRINT_EV("NUM_GPU_DEVICES", numGpuDevices,
@@ -722,6 +761,10 @@ public:
       printf("[AllToAll Related]\n");
     PRINT_EV("A2A_DIRECT", a2aDirect,
              std::string(a2aDirect ? "Only using direct links" : "Full all-to-all"));
+    PRINT_EV("A2A_MODE", a2aMode,
+             std::string(a2aMode == 0 ? "Perform copy" :
+                         a2aMode == 1 ? "Perform read-only" :
+                                        "Perform write-only"));
     PRINT_EV("USE_FINE_GRAIN", useFineGrain,
              std::string("Using ") + (useFineGrain ? "fine" : "coarse") + "-grained memory");
     PRINT_EV("USE_REMOTE_READ", useRemoteRead,